# swagger-api-2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "shop-food-api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/wh/inventory/create": {
      "post": {
        "tags": [
          "InventoryWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryWhCreateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryWhCreateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryWhCreateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryWhCreateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryWhCreateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryWhCreateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/inventory/update": {
      "post": {
        "tags": [
          "InventoryWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryWhUpdateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryWhUpdateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryWhUpdateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryWhUpdateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryWhUpdateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryWhUpdateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/inventory/delete": {
      "post": {
        "tags": [
          "InventoryWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryWhDeleteModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryWhDeleteModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryWhDeleteModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryWhDeleteModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryWhDeleteModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryWhDeleteModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/inventory/list": {
      "post": {
        "tags": [
          "InventoryWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryWhListModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryWhListModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryWhListModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryWhListModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryWhListModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryWhListModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/product/create": {
      "post": {
        "tags": [
          "ProductWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductWhCreateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductWhCreateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductWhCreateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWhCreateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWhCreateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWhCreateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/product/update": {
      "post": {
        "tags": [
          "ProductWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductWhUpdateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductWhUpdateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductWhUpdateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWhUpdateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWhUpdateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWhUpdateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/product/delete": {
      "post": {
        "tags": [
          "ProductWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductWhDeleteModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductWhDeleteModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductWhDeleteModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWhDeleteModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWhDeleteModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWhDeleteModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/product/list": {
      "post": {
        "tags": [
          "ProductWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductWhListModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductWhListModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductWhListModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWhListModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWhListModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductWhListModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/wh/warehouse/create": {
      "post": {
        "tags": [
          "WarehouseWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseCreateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseCreateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseCreateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseCreateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseCreateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseCreateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/warehouse/update": {
      "post": {
        "tags": [
          "WarehouseWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseUpdateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseUpdateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseUpdateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseUpdateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseUpdateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseUpdateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/warehouse/delete": {
      "post": {
        "tags": [
          "WarehouseWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseDeleteModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseDeleteModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseDeleteModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseDeleteModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseDeleteModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseDeleteModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/warehouse/list": {
      "post": {
        "tags": [
          "WarehouseWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseListModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseListModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseListModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseListModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseListModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseListModelResApiResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Exception": {
        "type": "object",
        "properties": {
          "targetSite": {
            "$ref": "#/components/schemas/MethodBase"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "data": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true,
            "readOnly": true
          },
          "innerException": {
            "$ref": "#/components/schemas/Exception"
          },
          "helpLink": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "hResult": {
            "type": "integer",
            "format": "int32"
          },
          "stackTrace": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "InventoryWhCreateModelReq": {
        "type": "object",
        "additionalProperties": false
      },
      "InventoryWhCreateModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "InventoryWhCreateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/InventoryWhCreateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "InventoryWhDeleteModelReq": {
        "type": "object",
        "additionalProperties": false
      },
      "InventoryWhDeleteModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "InventoryWhDeleteModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/InventoryWhDeleteModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "InventoryWhListModelReq": {
        "type": "object",
        "additionalProperties": false
      },
      "InventoryWhListModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "InventoryWhListModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/InventoryWhListModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "InventoryWhUpdateModelReq": {
        "type": "object",
        "additionalProperties": false
      },
      "InventoryWhUpdateModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "InventoryWhUpdateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/InventoryWhUpdateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
...
      "MetaData": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "exceptionExtra": {
            "$ref": "#/components/schemas/Exception"
          }
        },
        "additionalProperties": false
      },
...
      "ProductWhCreateModelReq": {
        "type": "object",
        "additionalProperties": false
      },
      "ProductWhCreateModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "ProductWhCreateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProductWhCreateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "ProductWhDeleteModelReq": {
        "type": "object",
        "additionalProperties": false
      },
      "ProductWhDeleteModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "ProductWhDeleteModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProductWhDeleteModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "ProductWhListModelReq": {
        "type": "object",
        "additionalProperties": false
      },
      "ProductWhListModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "ProductWhListModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProductWhListModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "ProductWhUpdateModelReq": {
        "type": "object",
        "additionalProperties": false
      },
      "ProductWhUpdateModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "ProductWhUpdateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProductWhUpdateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
...
      "WarehouseCreateModelReq": {
        "required": [
          "address",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "address": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "WarehouseCreateModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "WarehouseCreateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WarehouseCreateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "WarehouseDeleteModelReq": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "WarehouseDeleteModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "WarehouseDeleteModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WarehouseDeleteModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "WarehouseInfoModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WarehouseListModelReq": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "WarehouseListModelRes": {
        "type": "object",
        "properties": {
          "listWarehouse": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WarehouseInfoModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WarehouseListModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WarehouseListModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "WarehouseUpdateModelReq": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WarehouseUpdateModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "WarehouseUpdateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WarehouseUpdateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}